backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
selenium>=4.18,<5.0
beautifulsoup4>=4.12,<5.0
webdriver-manager>=4.0,<5.0
chromadb>=0.4,<1.0
openai>=1.12,<2.0
python-dotenv>=1.0,<2.0
tiktoken>=0.6,<1.0
pymysql>=1.1,<2.0
